packages: install debuginfo directly with debuginfo-install instead of dnf module
We had to explicitly enable list of debuginfo repositories when using dnf module.
This always caused trouble with rawhide which contains disabled fedora-* repositories
which does not work. So we either need to always special case rawhide and set there
special list of repositories until it becomes branched or we can use debuginfo-install
directly which enables needed repositories automatically.